--- src/main/java/com/notmarra/notcredits/commands/ReloadCommand.java
@@ -1,23 +1,0 @@
-package com.notmarra.notcredits.commands;
-
-import com.notmarra.notcredits.Notcredits;
-import com.notmarra.notcredits.utilities.Messages;
-import org.bukkit.Bukkit;
-import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
-
-public class ReloadCommand {
-    public static void execute(CommandSender sender) {
-        if (sender instanceof Player p) {
-            if (!p.hasPermission("notcredits.reload")) {
-                p.sendMessage(Messages.mm(Messages.messageGetString("no-perm")));
-            } else {
-                Notcredits.getInstance().reload();
-                p.sendMessage(Messages.mm(Messages.messageGetString("reload")));
-            }
-        } else {
-            Notcredits.getInstance().reload();
-        }
-
-    }
-}
