--- src/main/java/com/notmarra/notcredits/utilities/CheckVersion.java
@@ -1,8 +1,0 @@
-package com.notmarra.notcredits.utilities;
-
-public class CheckVersion {
-    public static boolean isSpigot() {
-        String version = org.bukkit.Bukkit.getServer().getVersion();
-        return version.toLowerCase().contains("spigot");
-    }
-}
