--- src/main/java/com/notmarra/notcredits/utilities/Decimal.java
@@ -1,17 +1,0 @@
-package com.notmarra.notcredits.utilities;
-
-import com.notmarra.notcredits.Notcredits;
-
-import java.text.DecimalFormat;
-
-public class Decimal {
-    public static String formatBalance(double value) {
-        if (Notcredits.getInstance().getConfig().getBoolean("balance_decimal")) {
-            DecimalFormat decimalFormat = new DecimalFormat(Notcredits.getInstance().getConfig().getString("balance_format"));
-            return String.valueOf(Double.parseDouble(decimalFormat.format(value)));
-        } else {
-            return String.valueOf(Math.round(value));
-        }
-    }
-
-}
